refactor: consolidate all list method cursor tests into one file
- Move list_tools test into test_list_methods_cursor.py
- Remove redundant test_list_tools_cursor.py file
- All cursor pagination tests are now in one place
