--- tests/client/test_list_tools_cursor.py
@@ -1,46 +1,0 @@
-import pytest
-
-from mcp.server.fastmcp import FastMCP
-from mcp.shared.memory import (
-    create_connected_server_and_client_session as create_session,
-)
-
-# Mark the whole module for async tests
-pytestmark = pytest.mark.anyio
-
-
-async def test_list_tools_cursor_parameter():
-    """Test that the cursor parameter is accepted in various forms.
-    
-    Note: FastMCP doesn't currently implement pagination, so these tests
-    only verify that the cursor parameter is accepted by the client.
-    """
-    server = FastMCP("test")
-
-    # Create a couple of test tools
-    @server.tool(name="test_tool_1")
-    async def test_tool_1() -> str:
-        """First test tool"""
-        return "Result 1"
-
-    @server.tool(name="test_tool_2")
-    async def test_tool_2() -> str:
-        """Second test tool"""
-        return "Result 2"
-
-    async with create_session(server._mcp_server) as client_session:
-        # Test without cursor parameter (omitted)
-        result1 = await client_session.list_tools()
-        assert len(result1.tools) == 2
-
-        # Test with cursor=None
-        result2 = await client_session.list_tools(cursor=None)
-        assert len(result2.tools) == 2
-
-        # Test with cursor as string
-        result3 = await client_session.list_tools(cursor="some_cursor_value")
-        assert len(result3.tools) == 2
-
-        # Test with empty string cursor
-        result4 = await client_session.list_tools(cursor="")
-        assert len(result4.tools) == 2
